Removed the blur on Remaincalm plugins' logo.
The screenshot is slightly blurred. A blur of 1px, which appears to
be the minimum, is too much compared to the screenshot, but not blur is too sharp.

I left 0.5px, which in my screen has no effect, hoping that it might be more accurate for
someone. If it's rendered as 1px, result is pleasant anyway.